Add missing showNullError updates
There is a conditional `showNullError` section in ForgotPassword.js, but since `showNullError` is never updated, I don't think this would ever be shown. I've added the necessary state updates to the `setState` calls and the constructor. Haven't tested whether this works (wrote it in the GitHub editor), but I don't have any reason to believe that it wouldn't :)